Moved inner class Network.Message into its own file
The class is not directly related to Network (could have been static) and is also used in other classes.
